requirements.txt
fastapi
pydantic
uvicorn
starlette
sqlalchemy
loguru
bcrypt
passlib
psycopg2
paho-mqtt